Cleared all warnings during make. Added comments for hint related function. Removed uneccessary functions and couts

+++ levelThree.cc
@@ -8,7 +8,7 @@ LevelThree::LevelThree(Board* board): Level(board) {
 // get the next block on this level
 std::unique_ptr<Block> LevelThree::getNext() {
   if (norandom) {
-    if (current == sequence.size()) {
+    if (current == (int) sequence.size()) {
       current = 0;
     }
     return Block::create(sequence[current++], board->getCurrentLevel());
